[EI-192] [web] refactor: Refactor hoverRender prop to ExpandableListItem component
