+++ web/app/ui/components/view/molocule/expandable-list-item/expandable-list-item.stories.tsx
@@ -1,6 +1,9 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
 import ExpandableListItem from './expandable-list-item';
+import { useState } from 'react';
+import IconButton from '../../atom/icons/icon-button';
+import { DotsHorizontalIcon } from '@radix-ui/react-icons';
 
 const meta = {
   title: 'view/molecule/ExpandableListItem',
@@ -17,4 +20,30 @@ const meta = {
 export default meta;
 type Story = StoryObj<typeof meta>;
 
-export const Default: Story = {};
+export const Default = {
+  render: () => {
+    const [selected, setSelected] = useState(false);
+
+    const hoverRender = () => {
+      return (
+        <IconButton
+          aria-label="edit"
+          // ref={iconButtonRef}
+          // onClick={handleIconButton}
+          icon={DotsHorizontalIcon}
+          color={'violet'}
+          className="mr-5"
+        />
+      );
+    };
+
+    return (
+      <ExpandableListItem
+        selected={selected}
+        onSelect={() => setSelected(true)}
+        onDeSelect={() => setSelected(false)}
+        hoverRender={hoverRender}
+      />
+    );
+  },
+};
